Improve input sanitization
Previously, sanitization of reactants before global featurization was inconsistent.
E.g. in the OneHotEncoder, reactants would be canonicalized and stipped of salts and have monomer protecting group chirality removed.
In other featurizers, inputs would only be canonicalized.

Now, we move all sanitization into the dataloaders. All global featurizers will receive the exact same input.
Sanitization is generally more rigorous to prepare the models for production.
All reactants will be stripped of counterions/desalted, so the models are invariant to receiving e.g. free base or HCl salt of the terminator as input.
The monomer protecting group chirality will always be removed. This is b/c the configuration can not be inferred from the product, is likely irrelevant for reactivity and is unavailable to most models anyway.

Other improvements:
- tests to cover the new functionality
- minor fix of a test that would fail under an unlikely random condition